+++ src/docindex.ts
@@ -54,7 +54,6 @@ import Tween, {
     CubicInOutTween, CubicInTween, CubicOutTween,
     QuarticInOutTween, QuarticInTween, QuarticOutTween,
     MultiTween, StepsTween } from './Tweens';
-import Updater from './Updater';
 import WebAudioMidiOut, { WebAudioMidiOutChannel } from './WebAudioMidiOut';
 
 export {
@@ -106,6 +105,5 @@ export {
         QuadraticInOutTween, QuadraticInTween, QuadraticOutTween,
         CubicInOutTween, CubicInTween, CubicOutTween,
         QuarticInOutTween, QuarticInTween, QuarticOutTween,
-    Updater,
     WebAudioMidiOut, WebAudioMidiOutChannel
 }